Waf patch for Fortran .mod dependencies
This patch makes Fortran source nodes that USE a module depend only on
the .mod node and not on the object file generated from the module
source file. This makes it possible to reduce recompilation when the
module source file changes in a trivial way that doesn't affect the
contents of the .mod file.

In other words, if foo.f has "USE mymod", and mymod is defined in
mymod.f, we turn this dependency graph:

  foo.f ---> mymod.mod ---> mymod.f
   |                        ^
   |                        |
   \-------> mymod.o -------/

Into this:

  foo.f ---> mymod.mod ---> mymod.f
                            ^
                            |
             mymod.o -------/

In practice, when using ifort, to get the benefit described above one
also needs to do two more things:

1) Decorate the fc class with Task.update_outputs. Not done as part of
this patch to leave the decision up to the wscript author.

2) Monkey-patch Utils.h_file() to strip the silly timestamp that ifort
inserts in the .mod files it writes, but that's not included in this
patch because it is an implementation-dependent hack (this patch fixes
an actual bug IMHO). Note: even stripping the timestamp is not perfect
because ifort sometimes inserts other mystery junk into the .mod file,
but at least we can avoid unnecessary recompilation *some* of the time.

If 1) is not done, this patch won't change the current behavior, which
is to have foo.f effectively be dependent on mymod.f because mymod.mod
inherits the signature from mymod.f.