python: use PREFIX in install directory if it is set
As a regression / misbehaviour compared to waf 2.0 when a Python
module is installed it seems that, even tho `PREFIX` is set for the whole
project, it is ignored.

In the past this was not ignored, namely:

```
(pydir,) = conf.get_python_variables(["get_python_lib(standard_lib=0, prefix=%r) or ''" % conf.env.PREFIX])
```

While with the current version we would always get (if not forced
with specific `PYTHONDIR` and `PYTHONARCHDIR`) the hardcoded path defined
at compilation time (`/usr/local/lib` in all cases I saw)

This patch takes this into account, using the user schema that permits to
customize the output passing the userbase variable, which is set to `PREFIX`
value indeed.

sysconfig.get_preferred_scheme('user') could give us the user schema, but
this is Python >= 3.10 only, so instead we construct the name manually
depending on the OS.