qt5: Detect the mkspec in use
The mkspec folder is always passed as an include path with each library.
In each mkspec folder is a file called qplatformdefs.h which contains
necessary includes and defines needed to build qt for that platform.

I'm not sure if the user is supposed to be able to override this, but
for now some basic override functionality is there when pkg-config is
not used. I've asked the Qt maintainers, but haven't gotten an answer on
this yet.