Support MSVC side-by-side (SxS) installs
Provides an extension to the msvc tool to allow for Waf projects
to configure against side-by-side installations of Visual Studio
and the Windows SDK. The extension implements discovery in a naive
method by scraping the toolset version from directory names in
existing installations identified by vswhere. This process was
simpler comparatively to trying to enumerate the component ids in
other fashions as the ids are not well documented and have been
inconsistently named over the versions supported (see re_sxs_ids).

See:
  https://learn.microsoft.com/en-us/visualstudio/install/install-visual-studio-versions-side-by-side?view=vs-2022
  https://learn.microsoft.com/en-us/visualstudio/install/tools-for-managing-visual-studio-instances?view=vs-2022